+++ src/App.tsx
@@ -1,6 +1,6 @@
 import "./App.css";
-import ATMContainer from "./features/atm/AtmContainer";
-import { AtmProvider } from "./features/atm/ATMContext";
+import ATMContainer from "./features/atm/components/AtmContainer";
+import { AtmProvider } from "./features/atm/store/ATMContext";
 
 function App() {
   return (
