+++ src/features/atm/services/Atm.service.ts
@@ -1,6 +1,6 @@
 import axios, { AxiosResponse } from "axios";
 
-import { API_ENDPOINTS } from "../../config/config";
+import { API_ENDPOINTS } from "../../../config/config";
 
 const instance = axios.create({
   baseURL: API_ENDPOINTS.base,
