+++ src/features/atm/types.d.ts
@@ -1,3 +1,8 @@
+type AtmContextProps = {
+  state: AtmState;
+  dispatch: Dispatch<AtmAction>;
+};
+
 type AtmState = {
   step: AtmSteps;
   pin: string;
